+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.0] - 2015-08-04
+### Added
+- `boolean` formatting helper, will show 'Yes' or 'No' depending on
+  attribute value.
+- Install generator. Copies locales into project.
+
 ## [0.1.1] - 2015-08-04
 ### Fixed
 - Disabeling label with format helpers shows unescaped HTML.